done tests and refactor

+++ api/src/app/controllers/InterpretationController.js
@@ -1,6 +1,6 @@
 import Interpretation from "../models/Interpretation";
 
-import { isAdmin } from "../utils";
+import { isAdmin } from "./helpers";
 
 export class InterpretationController {
   static async index(req, res) {
@@ -24,69 +24,48 @@ export class InterpretationController {
       ],
     };
 
-    try {
-      const interpretations = await Interpretation.paginate(options);
+    const interpretations = await Interpretation.paginate(options);
 
-      return res.status(200).json(interpretations);
-    } catch {
-      return res.status(500).json({
-        error: 500,
-        message: "Error on list interpretations.",
-      });
-    }
+    return res.status(200).json(interpretations);
   }
 
   static async store(req, res) {
     const { userId } = req;
     const { interpretation, music } = req.body;
 
-    try {
-      const model = await Interpretation.create({
-        interpretation,
-        music_id: music,
-        author_id: userId,
-      });
+    const model = await Interpretation.create({
+      interpretation,
+      music_id: music,
+      author_id: userId,
+    });
 
-      return res.status(201).json({
-        interpretation: model,
-      });
-    } catch {
-      return res.status(500).json({
-        error: 500,
-        message: "Error on create interpretation.",
-      });
-    }
+    return res.status(201).json({
+      interpretation: model,
+    });
   }
 
   static async show(req, res) {
     const { id } = req.params;
 
-    try {
-      const interpretation = await Interpretation.findByPk(id, {
-        include: [
-          {
-            association: "music",
-            include: {
-              association: "category",
-            },
+    const interpretation = await Interpretation.findByPk(id, {
+      include: [
+        {
+          association: "music",
+          include: {
+            association: "category",
           },
-        ],
-      });
+        },
+      ],
+    });
 
-      if (!interpretation) {
-        return res.status(404).json({
-          error: 404,
-          message: "Interpretation not found.",
-        });
-      }
-
-      return res.status(200).json({ interpretation });
-    } catch {
-      return res.status(500).json({
-        error: 500,
-        message: "Error on show interpretation.",
+    if (!interpretation) {
+      return res.status(404).json({
+        error: 404,
+        message: "Interpretation not found.",
       });
     }
+
+    return res.status(200).json({ interpretation });
   }
 
   static async update(req, res) {
@@ -94,69 +73,55 @@ export class InterpretationController {
     const { userId } = req;
     const { interpretation, music } = req.body;
 
-    try {
-      const model = await Interpretation.findByPk(id);
+    const model = await Interpretation.findByPk(id);
 
-      if (!model) {
-        return res.status(404).json({
-          error: 404,
-          message: "Not found interpretation.",
-        });
-      }
-
-      if (model.author_id !== userId && !isAdmin(userId)) {
-        return res.status(403).json({
-          error: 403,
-          message: "You not have access this resource.",
-        });
-      }
-
-      model.music_id = music;
-      model.interpretation = interpretation;
-      await model.save();
-
-      return res.status(200).json({
-        interpretation: model,
-      });
-    } catch {
-      return res.status(500).json({
-        error: 500,
-        message: "Error on update interpretation.",
+    if (!model) {
+      return res.status(404).json({
+        error: 404,
+        message: "Not found interpretation.",
       });
     }
+
+    if (model.author_id !== userId || !isAdmin(userId)) {
+      return res.status(403).json({
+        error: 403,
+        message: "You not have access this resource.",
+      });
+    }
+
+    model.music_id = music;
+    model.interpretation = interpretation;
+    await model.save();
+
+    return res.status(200).json({
+      interpretation: model,
+    });
   }
 
   static async destroy(req, res) {
     const { id } = req.params;
     const { userId } = req;
 
-    try {
-      const interpretation = await Interpretation.findByPk(id);
+    const interpretation = await Interpretation.findByPk(id);
 
-      if (!interpretation) {
-        return res.status(404).json({
-          error: 404,
-          message: "Interpretation not found.",
-        });
-      }
-
-      if (interpretation.author_id !== userId && !isAdmin(userId)) {
-        return res.status(403).json({
-          error: 403,
-          message: "You not have access this resource.",
-        });
-      }
-
-      await Interpretation.destroy({
-        where: { id },
-      });
-
-      return res.status(204).end();
-    } catch {
-      return res.status().json({
-        error: 500,
-        message: "Error on destroy interpretation.",
+    if (!interpretation) {
+      return res.status(404).json({
+        error: 404,
+        message: "Interpretation not found.",
       });
     }
+
+    if (interpretation.author_id !== userId || !isAdmin(userId)) {
+      return res.status(403).json({
+        error: 403,
+        message: "You not have access this resource.",
+      });
+    }
+
+    await Interpretation.destroy({
+      where: { id },
+    });
+
+    return res.status(204).end();
   }
 }
